Move Home and FilteredStations to containers
* Move Home and FilteredStations to containers
* Closes #32
